style: rename folders
