+++ src/components/Avatar/__tests__/Avatar.test.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import Avatar from '../Avatar';
 
 describe('Avatar', () => {
-  it('Shows region', () => {
+  it('Render component', () => {
     render(<Avatar src="test" size="md" />);
 
     const testImage = document.querySelector('img') as HTMLImageElement;
