+++ src/components/LoadingSpinner/index.ts
@@ -1,0 +1,3 @@
+import LoadingSpinner from './LoadingSpinner';
+
+export default LoadingSpinner;
