+++ src/components/PageMenu/index.ts
@@ -1,0 +1,3 @@
+import PageMenu from './PageMenu';
+
+export default PageMenu;
